Realm Online Status Rework
o Realm status is no longer defined by a timespan in which they ping the
Authorization Server. The ping will (in a future commit) help the auth
server know which RemoteIdentifier to use for the realm in question, and
base the status off if whether this ID is connected or not.